Make user script idempotent

+++ user-run.sh
@@ -13,56 +13,36 @@ email=$1
 name=$2
 
 # Set up aliases and other shell configurations
-echo '# Aliases' >> ~/.bash_aliases
-echo 'alias c="clear"' >> ~/.bash_aliases
-echo 'alias gs="git status"' >> ~/.bash_aliases
-echo 'alias ga="git add"' >> ~/.bash_aliases
-echo 'alias gd="git diff"' >> ~/.bash_aliases
-echo 'alias gc="git commit"' >> ~/.bash_aliases
-echo 'alias gl="git lg"' >> ~/.bash_aliases
-echo 'alias gpom="git push origin master"' >> ~/.bash_aliases
-echo 'alias gpr="git pull --rebase"' >> ~/.bash_aliases
-echo 'alias kb="kubectl"' >> ~/.bash_aliases
-echo 'alias kgp="kubectl get pods"' >> ~/.bash_aliases
-echo '' >> ~/.bash_aliases
-echo '# Configurations' >> ~/.bash_aliases
-echo 'set -o vi' >> ~/.bash_aliases
-echo '' >> ~/.bash_aliases
-echo '# Functions' >> ~/.bash_aliases
-echo '' >> ~/.bash_aliases
-echo '# Creates a new branch that tracks origin/master' >> ~/.bash_aliases
-echo 'function gcb {' >> ~/.bash_aliases
-echo '  local branch_name="$1"' >> ~/.bash_aliases
-echo '  git checkout -b "$branch_name"' >> ~/.bash_aliases
-echo '  git branch --set-upstream-to origin/master' >> ~/.bash_aliases
-echo '}' >> ~/.bash_aliases
-. ~/.bash_aliases
+cp ./customizations.bash ~/.customizations.bash
+if ! cat ~/.bashrc | grep customizations.bash &> /dev/null
+then
+  echo '. ~/.customizations.bash' >> ~/.bashrc
+fi
+. ~/.bashrc
 
 # Hide stuff I don't want in my home folder
-echo snap >> ~/.hidden
-echo Desktop >> ~/.hidden
-
-# Install jekyll for signed in user
-echo 'export GEM_HOME=$HOME/.gems' >> ~/.bashrc
-echo 'export PATH=$HOME/.gems/bin:$PATH' >> ~/.bashrc
-export GEM_HOME=$HOME/.gems
-export PATH=$HOME/.gems/bin:$PATH
-gem install jekyll bundler
+if ! cat ~/.hidden | grep snap &> /dev/null
+then
+  echo snap >> ~/.hidden
+fi
+if ! cat ~/.hidden | grep Desktop &> /dev/null
+then
+  echo Desktop >> ~/.hidden
+fi
 
 # Delete default directories
-rmdir ~/Documents || true
-rmdir ~/Downloads || true
-rmdir ~/Music || true
-rmdir ~/Pictures || true
-rmdir ~/Public || true
-rmdir ~/Templates || true
-rmdir ~/Videos || true
+rmdir ~/Documents &> /dev/null || true
+rmdir ~/Downloads &> /dev/null || true
+rmdir ~/Music &> /dev/null || true
+rmdir ~/Pictures &> /dev/null || true
+rmdir ~/Public &> /dev/null || true
+rmdir ~/Templates &> /dev/null || true
+rmdir ~/Videos &> /dev/null || true
 
 # Delete default directories from bookmarks
 truncate -s 0 ~/.config/user-dirs.dirs
 
-# This is necessary to set gsettings from within a script. Otherwise it doesn't
-# work
+# This is necessary to set gsettings from within a script. Otherwise it doesn't work
 export XDG_RUNTIME_DIR=/run/user/$myid
 export DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/$myid/bus
 
@@ -103,23 +83,47 @@ gsettings set org.gnome.Terminal.Legacy.Profile:/org/gnome/terminal/legacy/profi
 gsettings set org.gnome.Terminal.Legacy.Profile:/org/gnome/terminal/legacy/profiles:/:$GNOME_TERMINAL_PROFILE/ foreground-color '#AFAFAF'
 
 # Create an ssh key pair for github
-ssh-keygen -t rsa -b 4096 -C "$email" -f ~/.ssh/github_rsa -N ""
-eval "$(ssh-agent -s)"
-ssh-add ~/.ssh/github_rsa
+if test -f ~/.ssh/github_rsa
+then
+  echo "Github ssh key pair already exists"
+else
+  echo "Creating github ssh key pair"
+  ssh-keygen -t rsa -b 4096 -C "$email" -f ~/.ssh/github_rsa -N ""
+  eval "$(ssh-agent -s)"
+  ssh-add ~/.ssh/github_rsa
+fi
 
 # Configure git
 git config --global core.editor vim
 git config --global alias.lg "log --graph --all --pretty=format:'%Cred%h -%C(yellow)%d%Creset %s %Cgreen(%ci) %C(bold blue)<%an>'"
 git config --global user.email "$email"
-git config --global user.name "Adrian Ancona Novelo"
+git config --global user.name "$name"
 
 # Install vim
-mkdir -p ~/repos/
-git clone https://github.com/soonick/get-vim.git $HOME/repos/get-vim
-~/repos/get-vim/do.sh <<< "$HOME/bin/vim"
+if command -v nvim &> /dev/null
+then
+  echo "Vim is already installed"
+else
+  echo "Installing vim"
+  mkdir -p ~/repos/
+
+  if ! test -d $HOME/repos/get-vim
+  then
+    git clone https://github.com/soonick/get-vim.git $HOME/repos/get-vim
+  fi
+  current_folder=$(pwd)
+  cd ~/repos/get-vim/
+  ./do.sh <<< "$HOME/bin/vim"
+  cd $current_folder
+fi
 
 # Generate GPG key
-cat >foo <<EOF
+if test -f ~/.ssh/gpg.pub
+then
+  echo "GPG key already exists"
+else
+  echo "Creating GPG key"
+  cat >foo <<EOF
 %echo Generating GPG key
 Key-Type: RSA
 Key-Length: 3072
@@ -131,42 +135,25 @@ Name-Email: $email
 %commit
 %echo done
 EOF
-key=$(gpg --batch --gen-key foo 2>&1 | grep "revocation certificate stored as")
-# Remove everything before last /
-key=$(echo $key | sed -e 's/^.*\///g')
-# Remove everything after the .
-key=$(echo $key | cut -d "." -f 1)
-gpg --armor --export $key > ~/.ssh/gpg.pub
-echo "Public key copied to ~/.ssh/gpg.pub"
-git config --global commit.gpgsign true
-git config --global user.signingkey $key
+  key=$(gpg --batch --gen-key foo 2>&1 | grep "revocation certificate stored as")
+  # Remove everything before last /
+  key=$(echo $key | sed -e 's/^.*\///g')
+  # Remove everything after the .
+  key=$(echo $key | cut -d "." -f 1)
+  gpg --armor --export $key > ~/.ssh/gpg.pub
+  echo "Public key copied to ~/.ssh/gpg.pub"
+  git config --global commit.gpgsign true
+  git config --global user.signingkey $key
+fi
 
-# Tmux configuration
-cat >~/.tmux.conf <<EOF
-# Change prefix key from Ctrl+b to Ctrl+a
-set -g prefix C-a
-
-# "Prefix" "r" reloads configuration file
-bind r source-file ~/.tmux.conf \; display-message "Tmux configuration reloaded"
-
-# Make it possible to scroll window context with the mouse
-set -g mouse on
-
-# Use 256 colors
-set -g default-terminal "xterm-256color"
-
-# Enable vim navigation in copy mode
-setw -g mode-keys vi
-
-# Use v to trigger selection
-bind-key -T copy-mode-vi v send-keys -X begin-selection
-
-# Use y to yank current selection and put it in the OS' clipboard
-bind-key -T copy-mode-vi y send -X copy-pipe-and-cancel "xclip -sel clip -i"
-
-# Increase scroll buffer
-set -g history-limit 20000
-EOF
+# # Tmux configuration
+if test -f ~/.tmux.conf
+then
+  echo ".tmux.conf already exists"
+else
+  echo "Creating tmux.conf"
+  cp ./tmux.conf ~/.tmux.conf
+fi
 
 # Things that couldn't be automated
 echo ""
